4.21*: Blog List Expansion, step 9 Changed the delete blog operation so that a blog can be deleted only by the user who added it

+++ part4/blogList/controllers/blogs.js
@@ -37,6 +37,15 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+  if (user._id.toString() !== blog.user.toString()) {
+      return response.status(401).json({error: 'user invalid'})
+  }
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
